Move libbeat filebeat dependency to libbeat
Libbeat had a dependency on filebeat because of file rotation. This PR removes the dependency by moving the file rotation code to
