+++ filebeat/input/file/file_windows.go
@@ -5,8 +5,6 @@ import (
 	"os"
 	"reflect"
 	"syscall"
-
-	"github.com/elastic/beats/libbeat/logp"
 )
 
 type StateOS struct {
@@ -47,31 +45,6 @@ func (fs StateOS) IsSame(state StateOS) bool {
 
 func (fs StateOS) String() string {
 	return fmt.Sprintf("%d-%d-%d", fs.IdxHi, fs.IdxLo, fs.Vol)
-}
-
-// SafeFileRotate safely rotates an existing file under path and replaces it with the tempfile
-func SafeFileRotate(path, tempfile string) error {
-	old := path + ".old"
-	var e error
-
-	// In Windows, one cannot rename a file if the destination already exists, at least
-	// not with using the os.Rename function that Golang offers.
-	// This tries to move the existing file into an old file first and only do the
-	// move after that.
-	if e = os.Remove(old); e != nil {
-		logp.Debug("filecompare", "delete old: %v", e)
-		// ignore error in case old doesn't exit yet
-	}
-	if e = os.Rename(path, old); e != nil {
-		logp.Debug("filecompare", "rotate to old: %v", e)
-		// ignore error in case path doesn't exist
-	}
-
-	if e = os.Rename(tempfile, path); e != nil {
-		logp.Err("rotate: %v", e)
-		return e
-	}
-	return nil
 }
 
 // ReadOpen opens a file for reading only
